requirements.txt
isbnlib
python-barcode[images]
matplotlib
numpy